Move mustDecodeSmallPolynomialHex to helpers_test.go
The helper lived in ntru_test.go but most callers were in
codec_test.go and falcon1024_test.go. Co-locating it with the other
shared test helpers (mustDecodeHex, requireEqualWords) matches where
readers will look for it.
